repo/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.40
psycopg[binary]==3.2.13
alembic==1.15.2
pydantic-settings==2.8.1
python-dotenv==1.0.1
pytest==8.3.5
httpx==0.28.1
python-multipart==0.0.20
cryptography==43.0.3
